backend/requirements.txt
alembic==1.14.0
altgraph==0.17.4
appnope==0.1.4
astroid==3.3.8
asttokens==3.0.0
attrs==24.3.0
Authlib==1.4.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
codecov==2.1.13
coverage==7.6.10
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
docopt==0.6.2
execnet==2.1.1
executing==2.1.0
fakeredis==2.26.2
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
freezegun==1.5.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
macholib==1.16.3
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
ordered-set==4.1.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peewee==3.17.8
pexpect==4.9.0
pickleshare==0.7.5
pip-autoremove==0.10.0
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.10.1
pylint==3.3.3
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-docker==3.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruff==0.8.4
setuptools==75.6.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
yarg==0.1.9
zstandard==0.23.0
